Extract the repair command into separate class
